remove MemStat.Net from Sekiban.Core.DotNet
