+++ src/Sekiban.Core.DotNet/Sekiban.Core.DotNet.csproj
@@ -24,7 +24,6 @@
 
     <ItemGroup>
         <PackageReference Include="MediatR" Version="12.4.1"/>
-        <PackageReference Include="MemStat.Net" Version="0.1.4"/>
         <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="9.0.0"/>
         <PackageReference Include="Microsoft.Extensions.Configuration" Version="9.0.0"/>
         <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="9.0.0"/>
@@ -41,7 +40,4 @@
         <PackageReference Include="System.Runtime.InteropServices" Version="4.3.0"/>
     </ItemGroup>
 
-    <ItemGroup>
-        <ProjectReference Include="..\MemStat.Net\MemStat.Net.csproj"/>
-    </ItemGroup>
 </Project>
